Revert "feat: support cl100k_base tokenization and increase performance for GPT2 (#3897)"
This reverts commit 739fc228c658eb5baba6498b118096cb9159962e.
