+++ haystack/nodes/answer_generator/__init__.py
@@ -1,5 +1,3 @@
-from haystack.utils.import_utils import safe_import
-
 from haystack.nodes.answer_generator.base import BaseGenerator
 from haystack.nodes.answer_generator.transformers import RAGenerator, Seq2SeqGenerator
 from haystack.nodes.answer_generator.openai import OpenAIAnswerGenerator
