+++ haystack/nodes/answer_generator/openai.py
@@ -1,26 +1,16 @@
+from typing import Optional, List, Tuple, Union
 import json
 import logging
-import sys
-from typing import List, Optional, Tuple, Union
-
 import requests
 
+from transformers import GPT2TokenizerFast
+
+from haystack.nodes.answer_generator import BaseGenerator
 from haystack import Document
 from haystack.errors import OpenAIError, OpenAIRateLimitError
-from haystack.nodes.answer_generator import BaseGenerator
 from haystack.utils.reflection import retry_with_exponential_backoff
 
 logger = logging.getLogger(__name__)
-
-USE_TIKTOKEN = False
-if sys.version_info >= (3, 8):
-    USE_TIKTOKEN = True
-
-if USE_TIKTOKEN:
-    import tiktoken  # pylint: disable=import-error
-else:
-    logger.warning("OpenAI tiktoken module is not available for Python < 3.8. Falling back to GPT2TokenizerFast.")
-    from transformers import GPT2TokenizerFast, PreTrainedTokenizerFast
 
 
 class OpenAIAnswerGenerator(BaseGenerator):
@@ -95,21 +85,12 @@ class OpenAIAnswerGenerator(BaseGenerator):
         self.examples_context = examples_context
         self.examples = examples
         self.stop_words = stop_words
+        self._tokenizer = GPT2TokenizerFast.from_pretrained("gpt2")
 
-        tokenizer = "gpt2"
         if "davinci" in self.model:
             self.MAX_TOKENS_LIMIT = 4000
-            if self.model.endswith("-003") and USE_TIKTOKEN:
-                tokenizer = "cl100k_base"
         else:
             self.MAX_TOKENS_LIMIT = 2048
-
-        if USE_TIKTOKEN:
-            logger.debug("Using tiktoken %s tokenizer", tokenizer)
-            self._tk_tokenizer: tiktoken.Encoding = tiktoken.get_encoding(tokenizer)
-        else:
-            logger.debug("Using GPT2TokenizerFast")
-            self._hf_tokenizer: PreTrainedTokenizerFast = GPT2TokenizerFast.from_pretrained(tokenizer)
 
     @retry_with_exponential_backoff(backoff_in_seconds=10, max_retries=5)
     def predict(
@@ -198,13 +179,8 @@ class OpenAIAnswerGenerator(BaseGenerator):
 
         qa_prompt = f"Q: {query}\nA:"
 
-        n_instruction_tokens = self._count_tokens(instruction + qa_prompt + "===\nContext: \n===\n")
-
-        logger.debug("Number of tokens in instruction: %s", n_instruction_tokens)
-
-        n_docs_tokens = [self._count_tokens(doc.content) for doc in documents]
-        logger.debug("Number of tokens in documents: %s", n_docs_tokens)
-
+        n_instruction_tokens = len(self._tokenizer.encode(instruction + qa_prompt + "===\nContext: \n===\n"))
+        n_docs_tokens = [len(self._tokenizer.encode(doc.content)) for doc in documents]
         # for length restrictions of prompt see: https://beta.openai.com/docs/api-reference/completions/create#completions/create-max_tokens
         leftover_token_len = self.MAX_TOKENS_LIMIT - n_instruction_tokens - self.max_tokens
 
@@ -238,12 +214,5 @@ class OpenAIAnswerGenerator(BaseGenerator):
         context = f"===\nContext: {context}\n===\n"
 
         full_prompt = instruction + context + qa_prompt
-        logger.debug("Full prompt: %s", full_prompt)
 
         return full_prompt, input_docs
-
-    def _count_tokens(self, text: str) -> int:
-        if USE_TIKTOKEN:
-            return len(self._tk_tokenizer.encode(text))
-        else:
-            return len(self._hf_tokenizer.tokenize(text))
